+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 0.15.5
+
 - Fix: UserTokenServer.delete wasn't using the correct key in to the :ets table
 - New: UserTokenServer.delete_for_shop/1 to delete all the tokens for a shop
 - New: Shop helper functions to translate from slug <> myshopify domain, and new to_uri helper function.